requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlmodel==0.0.24
psycopg2-binary==2.9.10
python-dotenv==1.1.1
python-multipart==0.0.20
email-validator==2.3.0
pydantic==2.11.7
pydantic-settings==2.10.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
